Address UI editor review follow-ups
Reviewer findings addressed:\n- Preserve unsaved UI editor drafts per surface when switching documents.\n- Require an admin bearer token for live client preview requests.\n- Move the UI layout schema into the shared @silencer/ui-layout TypeScript workspace and use it from admin-web and admin-api.\n\nAlso updated Docker workspace copies so both admin images include the shared package.
