+++ package.json
@@ -9,6 +9,7 @@
     "services/admin-api",
     "shared/fonts",
     "shared/gas-validation",
+    "shared/ui-layout",
     "web/admin",
     "web/website"
   ],
