+++ services/admin-api/package.json
@@ -7,6 +7,7 @@
     "dev": "bun --watch src/index.js"
   },
   "dependencies": {
+    "@silencer/ui-layout": "workspace:*",
     "amqplib": "^0.10.4",
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
