+++ web/admin/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@monaco-editor/react": "^4.7.0",
     "@silencer/gas-validation": "workspace:*",
+    "@silencer/ui-layout": "workspace:*",
     "@types/dagre": "^0.7.54",
     "dagre": "^0.8.5",
     "fflate": "^0.8.2",
